Remove RegisterPage component and integrate registration functionality into SignInPage
- Deleted the RegisterPage component to streamline the authentication process.
- Enhanced SignInPage to support both sign-in and registration, allowing users to toggle between modes.
- Added form fields for first name, last name, ethnicity, and nationality in the registration section.
- Implemented password validation and requirements display for improved user experience during registration.
- Updated CSS styles to accommodate the new layout and ensure a cohesive design across authentication pages.
